[util2.embed] move instagram-embed to util2
