+++ instagram_to_discord/sites/instagram/instagram.py
@@ -1,4 +1,6 @@
 from logging import getLogger,StreamHandler,INFO
+
+from instagram_to_discord.util2.embed import create_embed_instagram_image
 logger = getLogger(__name__)    #以降、このファイルでログが出たということがはっきりする。
 handler = StreamHandler()
 handler.setLevel(INFO)
@@ -13,10 +15,6 @@ from ...string_util import sophisticate_string
 from .converter_instagram_url import instagram_make_author_page
 from ...params import IS_DEBUG
 
-def create_embed_instagram_image(image_url: str):
-    embed = discord.Embed(color=discord.Color.red())
-    embed.set_image(url=image_url)
-    return embed
 
 async def send_instagram_images_from_cache_for_specified_index(
     skip_one: bool, image_urls: List[str], nums: List[int], message
@@ -55,33 +53,4 @@ async def send_instagram_images_for_specified_index(
         embed = create_embed_instagram_image(image_urls[idx])
         await message.channel.send(embed=embed)
 
-def create_instagram_pic_embed(obj: InstagramData, base_url: str):
-    description = sophisticate_string(obj.caption)
-    embed = discord.Embed(
-        title=obj.full_name,
-        description=description,
-        url=base_url,
-        color=discord.Color.red(),
-    )
-    embed.set_image(url=obj.media)
-    embed.set_author(
-        name=obj.full_name,
-        url=instagram_make_author_page(obj.username),
-        icon_url=obj.profile_url,
-    )
-    return embed
 
-def create_instagram_video_embed(obj: InstagramData, base_url: str):
-    description = sophisticate_string(obj.caption)
-    embed = discord.Embed(
-        title=obj.full_name,
-        description=description,
-        url=base_url,
-        color=discord.Color.red(),
-    )
-    embed.set_author(
-        name=obj.full_name,
-        url=instagram_make_author_page(obj.username),
-        icon_url=obj.profile_url,
-    )
-    return embed
