+++ instagram_to_discord/sites/instagram/instagram_process.py
@@ -1,4 +1,6 @@
 from logging import getLogger,StreamHandler,INFO,DEBUG
+
+from instagram_to_discord.util2.embed import create_instagram_pic_embed, create_instagram_video_embed
 logger = getLogger(__name__)    #以降、このファイルでログが出たということがはっきりする。
 handler = StreamHandler()
 handler.setLevel(INFO)
@@ -23,7 +25,7 @@ from .converter_instagram_url import (convert_instagram_url_to_a,
 from ...cookie_requests import requests_get_cookie
 from .instagram_type import (get_multiple_medias_from_str,
                              instagram_parse_json_to_obj)
-from .instagram import get_instagram_id_from_url, send_instagram_images_for_specified_index, create_instagram_pic_embed, create_instagram_video_embed
+from .instagram import get_instagram_id_from_url, send_instagram_images_for_specified_index
 from ...video import trimming_video_to_8MB
 
 async def process_instagram(client: Any, channel, message, content):
